requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
aiosqlite==0.22.1
anyio==4.13.0
asyncio-dgram==3.0.0
asyncpg==0.31.0
atlastk==0.13.5
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
discord.py==2.7.1
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
iso8601==2.1.0
mcstatus==13.0.1
multidict==6.7.1
numpy==2.4.4
pillow==12.2.0
propcache==0.4.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pypika-tortoise==0.6.5
python-dotenv==1.2.2
requests==2.33.1
tortoise==0.1.1
tortoise-orm==1.1.7
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0

apscheduler